+++ srcs/events/key_handle.c
@@ -6,51 +6,15 @@
 /*   By: hluiz-ma <hluiz-ma@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/30 21:20:50 by hluiz-ma          #+#    #+#             */
-/*   Updated: 2025/02/24 20:40:15 by marsoare         ###   ########.fr       */
+/*   Updated: 2025/03/09 12:12:34 by marsoare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include <cub3d.h>
 
-void	handle_movement(t_game *game)
+void	handle_movement_keys(int keycode, t_game *game)
 {
-	double	dir_x;
-	double	dir_y;
-
-	dir_x = 0;
-	dir_y = 0;
-	if (game->p1.keys.w)
-	{
-		dir_x += game->p1.dir.x;
-		dir_y += game->p1.dir.y;
-	}
-	if (game->p1.keys.s)
-	{
-		dir_x -= game->p1.dir.x;
-		dir_y -= game->p1.dir.y;
-	}
-	keys_else(game, &dir_x, &dir_y);
-	move_player(game, dir_x, dir_y);
-	if (game->p1.keys.left)
-		rotate_player(game, -game->p1.rot_speed);
-	if (game->p1.keys.right)
-		rotate_player(game, game->p1.rot_speed);
-	
-	// Handle shooting if player is firing
-	if (game->p1.is_firing)
-	{
-		shoot_enemy(game);
-		// Firing animation will handle resetting is_firing to 0
-	}
-}
-
-int	key_press(int keycode, t_game *game)
-{
-	printf("Key pressed: %d\n", keycode);
-	if (keycode == KEY_ESC)
-		close_window(game);
-	else if (keycode == KEY_E)
-		interact_with_door(game);
-	else if (keycode == KEY_W)
+	if (keycode == KEY_W)
 		game->p1.keys.w = 1;
 	else if (keycode == KEY_S)
 		game->p1.keys.s = 1;
@@ -62,6 +26,14 @@ int	key_press(int keycode, t_game *game)
 		game->p1.keys.left = 1;
 	else if (keycode == KEY_RIGHT)
 		game->p1.keys.right = 1;
+}
+
+void	handle_action_keys(int keycode, t_game *game)
+{
+	if (keycode == KEY_ESC)
+		close_window(game);
+	else if (keycode == KEY_E)
+		interact_with_door(game);
 	else if (keycode == KEY_SPACE || keycode == MOUSE_LEFT)
 	{
 		game->p1.is_firing = 1;
@@ -69,6 +41,12 @@ int	key_press(int keycode, t_game *game)
 		game->p1.last_fire = get_time_ms();
 		shoot_enemy(game);
 	}
+}
+
+int	key_press(int keycode, t_game *game)
+{
+	handle_movement_keys(keycode, game);
+	handle_action_keys(keycode, game);
 	return (0);
 }
